play next frame always no frame drop
Otherwise the result maybe wrong if video is slow and video will not
paused after a frame is displayed. set clock type to VideoClock can
solve the problem. But audio maybe too fast(video is too slow) and video
will skip many frames to sync to audio if normal playback is resumed.