static lib: register io&subtitle. fix vc link error
some components symbols in QtAV are not used by other files, for example
the AudioDecoderFFmpeg before 54a7f2cb. Their code can be removed by
linker (-dead_strip for ios) then the component can not be registered.
If we use the RegisterXXX_Man symbol in another unit that is stripped,
then that component will be linked to exe and registered (when the base
class is used! for example, subtile is not registered for a minimal
player, why?). But what about
the static construct var?