videotoolbox: try to support ios 0-copy
- iOS 0-copy is not finished. We need EAGL code in obj-c. The texture
ids generation is not controled by VideoShader. A possible solution is
map to textures in createHandle(), and move texture cache/ref to
decoder.
To avoid using EAGL, we can lock the cv buffer, get plane address and
upload to given textures. No ffmpeg modification is required
(kCVPixelBufferOpenGLESCompatibilityKey). But the performance is worse
than EAGL.
- Use rgb422_apple texture for uyvy422 to support both iOS and OSX