Copy over initial_padding, extradata_size and extradata from avctx to newly created audio stream
This is needed for audio codecs that for example use the CodecPrivate
element in the WEBM container. The CodecPrivate field can contain
headers or other additional data. The audio codec OPUS cannot be
decoded without this element. Some audio codecs also define an
initial_padding which can be written as CodecDelay into the WEBM
container.