Clearing packetqueue and setting clock from avdemux thread.
If the packetqueue was cleared in the seek function then it can happen that
seek clears the packetqueue, then the other thread puts an other packet
with high timestamp and videothread takes it. In this case the videothread
waits a lot and video just freezes. This was fixed by putting the clear
into the SeekTask's run function. When I fixed this I realized that the
clock is sometimes wrong,too, so I put the clock updates into the SeekTask,
too. I am not sure if this is the perfect solution to make videothread's
waitAndCheck always work correctly, but it's definitely improved.